+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 2.0.0
+
+Big rewrites
+
+-   Model select Popup
+-   Mark files for context
+-   Better system prompts
+-   More configuration values like temp
+
 # 1.1.0
 
 -   Added Telescope for picker.